backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
pymysql==1.1.0
SQLAlchemy==2.0.23
nltk==3.8.1
vaderSentiment==3.3.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
yfinance==0.2.31
PyJWT==2.8.0
matplotlib==3.8.0
plotly==5.17.0
pytest==7.4.2
gunicorn==21.2.0 